+++ docs/changelog/changelog.md
@@ -9,6 +9,7 @@
 ### Changes
 
 - On the `BaseLlmMap.process` the argument `choice_count` was renamed to `max_return_values` to make it more clear what it does and is now optional.
+- The `BaseListIntel` now requires you to set the type of the `items` class variable with a type hint.
 
 ## v0.17.0